requirements.txt
amqp==5.1.1
appnope==0.1.3
asgiref==3.7.1
astroid==2.15.5
asttokens==2.2.1
async-timeout==4.0.2
backcall==0.2.0
bandit==1.7.5
bidict==0.22.1
billiard==3.6.4.0
biosppy==1.0.0
black==23.3.0
celery==5.2.7
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
decorator==5.1.1
dill==0.3.6
Django==4.2.1
djangorestframework==3.14.0
executing==1.2.0
flake8==6.0.0
fonttools==4.39.4
future==0.18.3
gensim==4.3.1
gitdb==4.0.10
GitPython==3.1.31
gunicorn==20.1.0
h5py==3.8.0
idna==3.4
ipython==8.13.2
isort==5.12.0
jedi==0.18.2
joblib==1.2.0
kiwisolver==1.4.4
kombu==5.2.4
lazy-object-proxy==1.9.0
mando==0.7.1
markdown-it-py==2.2.0
matplotlib==3.7.1
matplotlib-inline==0.1.6
mccabe==0.7.0
mdurl==0.1.2
mypy-extensions==1.0.0
numpy==1.24.3
opencv-python==4.7.0.72
packaging==23.1
pandas==2.0.1
parso==0.8.3
pathspec==0.11.1
pbr==5.11.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.5.1
plotly==5.14.1
prompt-toolkit==3.0.38
psycopg2-binary==2.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.10.0
pycparser==2.21
pydeps==1.12.8
pyflakes==3.0.1
Pygments==2.15.1
pylint==2.17.4
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
radon==6.0.1
redis==4.5.5
requests==2.31.0
rich==13.4.1
scikit-learn==1.2.2
scipy==1.10.1
serializers==0.2.4
shortuuid==1.0.11
six==1.16.0
smart-open==6.3.0
smmap==5.0.0
soundfile==0.12.1
sqlparse==0.4.4
stack-data==0.6.2
stdlib-list==0.8.0
stevedore==5.1.0
tenacity==8.2.2
threadpoolctl==3.1.0
tk==0.1.0
tomli==2.0.1
tomlkit==0.11.8
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.6.1
tzdata==2023.3
urllib3==2.0.2
vine==5.0.0
wcwidth==0.2.6
wfdb==4.1.1
wrapt==1.15.0
